+++ go/libraries/doltcore/sqle/dsess/session.go
@@ -18,6 +18,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/dolthub/dolt/go/cmd/dolt/cli"
 	"strconv"
 	"strings"
 	"sync"
@@ -27,7 +28,6 @@ import (
 	sqltypes "github.com/dolthub/go-mysql-server/sql/types"
 	"github.com/shopspring/decimal"
 
-	"github.com/dolthub/dolt/go/cmd/dolt/cli"
 	"github.com/dolthub/dolt/go/libraries/doltcore/branch_control"
 	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb"
 	"github.com/dolthub/dolt/go/libraries/doltcore/env"
@@ -61,7 +61,7 @@ type DoltSession struct {
 	statsProv        sql.StatsProvider
 	mu               *sync.Mutex
 	fs               filesys.Filesys
-	writeSessProv    writeSessFunc
+	writeSessProv    WriteSessFunc
 
 	// If non-nil, this will be returned from ValidateSession.
 	// Used by sqle/cluster to put a session into a terminal err state.
@@ -74,7 +74,7 @@ var _ sql.TransactionSession = (*DoltSession)(nil)
 var _ branch_control.Context = (*DoltSession)(nil)
 
 // DefaultSession creates a DoltSession with default values
-func DefaultSession(pro DoltDatabaseProvider, sessFunc writeSessFunc) *DoltSession {
+func DefaultSession(pro DoltDatabaseProvider, sessFunc WriteSessFunc) *DoltSession {
 	return &DoltSession{
 		Session:          sql.NewBaseSession(),
 		username:         "",
@@ -91,8 +91,6 @@ func DefaultSession(pro DoltDatabaseProvider, sessFunc writeSessFunc) *DoltSessi
 	}
 }
 
-type writeSessFunc func(nbf *types.NomsBinFormat, ws *doltdb.WorkingSet, aiTracker globalstate.AutoIncrementTracker, opts editor.Options) WriteSession
-
 // NewDoltSession creates a DoltSession object from a standard sql.Session and 0 or more Database objects.
 func NewDoltSession(
 	sqlSess *sql.BaseSession,
@@ -100,7 +98,7 @@ func NewDoltSession(
 	conf config.ReadWriteConfig,
 	branchController *branch_control.Controller,
 	statsProvider sql.StatsProvider,
-	writeSessProv writeSessFunc,
+	writeSessProv WriteSessFunc,
 ) (*DoltSession, error) {
 	username := conf.GetStringOrDefault(config.UserNameKey, "")
 	email := conf.GetStringOrDefault(config.UserEmailKey, "")
@@ -1748,38 +1746,7 @@ func DefaultHead(baseName string, db SqlDatabase) (string, error) {
 	return head, nil
 }
 
-// WriteSession encapsulates writes made within a SQL session.
-// It's responsible for creating and managing the lifecycle of TableWriter's.
-type WriteSession interface {
-	// GetTableWriter creates a TableWriter and adds it to the WriteSession.
-	GetTableWriter(ctx *sql.Context, tableName doltdb.TableName, db string, setter SessionRootSetter) (TableWriter, error)
-
-	// SetWorkingSet modifies the state of the WriteSession. The WorkingSetRef of |ws| must match the existing Ref.
-	SetWorkingSet(ctx *sql.Context, ws *doltdb.WorkingSet) error
-
-	// GetOptions returns the editor.Options for this session.
-	GetOptions() editor.Options
-
-	// SetOptions sets the editor.Options for this session.
-	SetOptions(opts editor.Options)
-
-	WriteSessionFlusher
-}
-
-type TableWriter interface {
-	sql.TableEditor
-	sql.ForeignKeyEditor
-	sql.AutoIncrementSetter
-}
-
-// SessionRootSetter sets the root value for the session.
-type SessionRootSetter func(ctx *sql.Context, dbName string, root doltdb.RootValue) error
-
-// WriteSessionFlusher is responsible for flushing any pending edits to the session
-type WriteSessionFlusher interface {
-	// Flush flushes the pending writes in the session.
-	Flush(ctx *sql.Context) (*doltdb.WorkingSet, error)
-	// FlushWithAutoIncrementOverrides flushes the pending writes in the session, overriding the auto increment values
-	// for any tables provided in the map
-	FlushWithAutoIncrementOverrides(ctx *sql.Context, increment bool, autoIncrements map[string]uint64) (*doltdb.WorkingSet, error)
-}
+// WriteSessFunc is a constructor that session builders use to
+// create fresh table editors.
+// The indirection avoids a writer/dsess package import cycle.
+type WriteSessFunc func(nbf *types.NomsBinFormat, ws *doltdb.WorkingSet, aiTracker globalstate.AutoIncrementTracker, opts editor.Options) WriteSession
