+++ go/libraries/doltcore/sqle/dtables/docs_table.go
@@ -16,7 +16,6 @@ package dtables
 
 import (
 	"fmt"
-
 	"github.com/dolthub/go-mysql-server/sql"
 	sqlTypes "github.com/dolthub/go-mysql-server/sql/types"
 	"github.com/dolthub/vitess/go/sqltypes"
@@ -258,7 +257,12 @@ func (iw *docsWriter) StatementBegin(ctx *sql.Context) {
 	}
 
 	if ws := dbState.WriteSession(); ws != nil {
-		tableWriter, err := ws.GetTableWriter(ctx, doltdb.TableName{Name: doltdb.DocTableName}, dbName, dSess.SetWorkingRoot)
+		table, _, err := roots.Working.GetTable(ctx, doltdb.TableName{Name: doltdb.DocTableName})
+		if err != nil {
+			iw.errDuringStatementBegin = err
+			return
+		}
+		tableWriter, err := ws.GetTableWriter(ctx, table, doltdb.TableName{Name: doltdb.DocTableName}, dbName, dSess.SetWorkingRoot)
 		if err != nil {
 			iw.errDuringStatementBegin = err
 			return
